requirements.txt
asgiref==3.3.1
Django==3.1.4
django-allauth==0.44.0
django-crispy-forms==1.10.0
django-localflavor==3.0.1
django-mysql==3.9.0
mysql-connector==2.2.9
oauthlib==3.1.0
Pillow==8.0.1
PyJWT==1.7.1
python-stdnum==1.14
python3-openid==3.2.0
pytz==2020.4
requests-oauthlib==1.3.0
sqlparse==0.4.1
